core: pinctrl: fix review: remove sleep state, add zephyrPinctrl.cpp
