+++ variants/arduino_giga_r1_stm32h747xx_m7/arduino_giga_r1_stm32h747xx_m7.overlay
@@ -15,16 +15,14 @@
 	status = "okay";
 	/* usart1 is not set as deferred-init to catch early boot logs since it is assigned to zephyr,console */
 	pinctrl-0 = <&usart1_tx_pa9 &usart1_rx_pb7>;
-	pinctrl-1 = <&analog_pa9 &analog_pb7>;
-	pinctrl-names = "default", "sleep";
+	pinctrl-names = "default";
 };
 
 &usart2 {
 	status = "okay";
 	zephyr,deferred-init;
 	pinctrl-0 = <&usart2_tx_pd5 &usart2_rx_pd6>;
-	pinctrl-1 = <&analog_pd5 &analog_pd6>;
-	pinctrl-names = "default", "sleep";
+	pinctrl-names = "default";
 	current-speed = <115200>;
 };
 
@@ -32,8 +30,7 @@
 	status = "okay";
 	zephyr,deferred-init;
 	pinctrl-0 = <&uart4_tx_ph13 &uart4_rx_pi9>;
-	pinctrl-1 = <&analog_ph13 &analog_pi9>;
-	pinctrl-names = "default", "sleep";
+	pinctrl-names = "default";
 	current-speed = <115200>;
 };
 
@@ -41,8 +38,7 @@
 	status = "okay";
 	zephyr,deferred-init;
 	pinctrl-0 = <&usart6_tx_pg14 &usart6_rx_pc7>;
-	pinctrl-1 = <&analog_pg14 &analog_pc7>;
-	pinctrl-names = "default", "sleep";
+	pinctrl-names = "default";
 };
 
 &rtc {
@@ -73,8 +69,7 @@
 	status = "okay";
 	zephyr,deferred-init;
 	pinctrl-0 = <&i2c1_scl_pb8 &i2c1_sda_pb9>;
-	pinctrl-1 = <&analog_pb8 &analog_pb9>;
-	pinctrl-names = "default", "sleep";
+	pinctrl-names = "default";
 	clock-frequency = <I2C_BITRATE_FAST>;
 };
 
@@ -82,8 +77,7 @@
 	status = "okay";
 	zephyr,deferred-init;
 	pinctrl-0 = <&i2c2_scl_ph4 &i2c2_sda_pb11>;
-	pinctrl-1 = <&analog_ph4 &analog_pb11>;
-	pinctrl-names = "default", "sleep";
+	pinctrl-names = "default";
 	clock-frequency = <I2C_BITRATE_FAST>;
 };
 
@@ -95,7 +89,7 @@
 		status = "okay";
 		zephyr,deferred-init;
 		pinctrl-0 = <>;
-		pinctrl-1 = <&analog_pj9 &analog_pk1 &analog_pj11>;
+		pinctrl-1 = <>;
 		pinctrl-2 = <&tim1_ch3_pj9 &tim1_ch1_pk1 &tim1_ch2_pj11>;
 		pinctrl-names = "default", "sleep", "arduino";
 	};
@@ -109,8 +103,8 @@
 		status = "okay";
 		zephyr,deferred-init;
 		pinctrl-0 = <>;
-		pinctrl-1 = <&analog_pa3 &analog_pa2>;
-		pinctrl-2 = <&tim2_ch4_pa3  &tim2_ch3_pa2>;
+		pinctrl-1 = <>;
+		pinctrl-2 = <&tim2_ch4_pa3 &tim2_ch3_pa2>;
 		pinctrl-names = "default", "sleep", "arduino";
 	};
 };
@@ -123,7 +117,7 @@
 		status = "okay";
 		zephyr,deferred-init;
 		pinctrl-0 = <>;
-		pinctrl-1 = <&analog_pa7 &analog_pb4>;
+		pinctrl-1 = <>;
 		pinctrl-2 = <&tim3_ch2_pa7 &tim3_ch1_pb4>;
 		pinctrl-names = "default", "sleep", "arduino";
 	};
@@ -137,7 +131,7 @@
 		status = "okay";
 		zephyr,deferred-init;
 		pinctrl-0 = <>;
-		pinctrl-1 = <&analog_pd13 &analog_pb8 &analog_pb9>;	
+		pinctrl-1 = <>;	
 		pinctrl-2 = <&tim4_ch2_pd13 &tim4_ch3_pb8 &tim4_ch4_pb9>;
 		pinctrl-names = "default", "sleep", "arduino";
 	};
@@ -151,7 +145,7 @@
 		status = "okay";
 		zephyr,deferred-init;
 		pinctrl-0 = <>;
-		pinctrl-1 = <&analog_pj8 &analog_pj10>;
+		pinctrl-1 = <>;
 		pinctrl-2 = <&tim8_ch1_pj8 &tim8_ch2_pj10>;
 		pinctrl-names = "default", "sleep", "arduino";
 	};
@@ -165,7 +159,7 @@
 		status = "okay";
 		zephyr,deferred-init;
 		pinctrl-0 = <>;
-		pinctrl-1 = <&analog_ph6>;
+		pinctrl-1 = <>;
 		pinctrl-2 = <&tim12_ch1_ph6>;
 		pinctrl-names = "default", "sleep", "arduino";
 	};
@@ -224,8 +218,7 @@
 	zephyr,deferred-init;
 	pinctrl-0 = <&spi1_sck_pb3
 		     &spi1_miso_pg9 &spi1_mosi_pd7>;
-	pinctrl-1 = <&analog_pb3 &analog_pg9 &analog_pd7>;
-	pinctrl-names = "default", "sleep";
+	pinctrl-names = "default";
 };
 
 &spi5 {
@@ -233,8 +226,7 @@
 	zephyr,deferred-init;
 	pinctrl-0 = <&spi5_sck_ph6
 		     &spi5_miso_pj11 &spi5_mosi_pj10>;
-	pinctrl-1 = <&analog_ph6 &analog_pj11 &analog_pj10>;
-	pinctrl-names = "default", "sleep";
+	pinctrl-names = "default";
 };
 
 &adc1 {
@@ -352,8 +344,7 @@
 	zephyr,deferred-init;
 	pinctrl-0 = <>;
 	pinctrl-1 = <>;
-	pinctrl-2 = <&adc3_inp0_pc2_c
-		&adc3_inp1_pc3_c>;
+	pinctrl-2 = <&adc3_inp0_pc2_c &adc3_inp1_pc3_c>;
 	pinctrl-names = "default", "sleep", "arduino";
 	st,adc-clock-source = "SYNC";
 	st,adc-prescaler = <4>;
@@ -381,7 +372,7 @@
 &dac1 {
 	zephyr,deferred-init;
 	pinctrl-0 = <>;
-	pinctrl-1 = <&analog_pa4 &analog_pa5>;
+	pinctrl-1 = <>;
 	pinctrl-2 = <&dac1_out1_pa4 &dac1_out2_pa5>;
 	pinctrl-names = "default", "sleep", "arduino";
 };
