+++ variants/arduino_uno_q_stm32u585xx/arduino_uno_q_stm32u585xx.overlay
@@ -33,8 +33,7 @@
 	status = "okay";
 	zephyr,deferred-init;
 	pinctrl-0 = <&i2c3_scl_pc0 &i2c3_sda_pc1>;
-	pinctrl-1 = <&analog_pc0 &analog_pc1>;
-	pinctrl-names = "default", "sleep"; 
+	pinctrl-names = "default"; 
 	clock-frequency = <I2C_BITRATE_FAST>;
 };
 
@@ -111,7 +110,7 @@
 		/* tim1_etr_pa12 is not available for PWM */
 		zephyr,deferred-init;
 		pinctrl-0 = <>;
-		pinctrl-1 = <&analog_pa11 &analog_pb15 &analog_pb13 &analog_pb14>;
+		pinctrl-1 = <>;
 		pinctrl-2 = <&tim1_ch4_pa11 &tim1_ch3n_pb15 &tim1_ch1n_pb13 &tim1_ch2n_pb14>;
 		pinctrl-names = "default", "sleep", "arduino";
 	};
@@ -127,7 +126,7 @@
 		status = "okay";
 		zephyr,deferred-init;
 		pinctrl-0 = <>;
-		pinctrl-1 = <&analog_pb3 &analog_pb10 &analog_pb11>;
+		pinctrl-1 = <>;
 		pinctrl-2 = <&tim2_ch2_pb3 &tim2_ch3_pb10 &tim2_ch4_pb11>;
 		pinctrl-names = "default", "sleep", "arduino";
 	};
@@ -141,7 +140,7 @@
 		status = "okay";
 		zephyr,deferred-init;
 		pinctrl-0 = <>;
-		pinctrl-1 = <&analog_pb0 &analog_pb1 &analog_pb4>;
+		pinctrl-1 = <>;
 		pinctrl-2 = <&tim3_ch3_pb0 &tim3_ch4_pb1 &tim3_ch1_pb4>;
 		pinctrl-names = "default", "sleep", "arduino";
 	};
@@ -155,7 +154,7 @@
 		status = "okay";
 		zephyr,deferred-init;
 		pinctrl-0 = <>;
-		pinctrl-1 = <&analog_pb8 &analog_pb9 /*&analog_pb6 &analog_pb7*/>;
+		pinctrl-1 = <>;
 		pinctrl-2 = <&tim4_ch3_pb8 &tim4_ch4_pb9 /*&tim4_ch1_pb6 &tim4_ch2_pb7*/>;
 		pinctrl-names = "default", "sleep", "arduino";
 	};
@@ -169,7 +168,7 @@
         status = "okay";
         zephyr,deferred-init;
         pinctrl-0 = <>;
-        pinctrl-1 = <&analog_ph10 &analog_ph11 &analog_ph12>;
+        pinctrl-1 = <>;
         pinctrl-2 = <&tim5_ch1_ph10 &tim5_ch2_ph11 &tim5_ch3_ph12>;
         pinctrl-names = "default", "sleep", "arduino";
     };
@@ -178,7 +177,7 @@
 &adc1 {
 	zephyr,deferred-init;
 	pinctrl-0 = <>;
-	pinctrl-1 = <&analog_pa4 &analog_pa5 &analog_pa6 &analog_pa7 &analog_pc1 &analog_pc0>;
+	pinctrl-1 = <>;
 	pinctrl-2 = <&adc1_in9_pa4 &adc1_in10_pa5 &adc1_in11_pa6 &adc1_in12_pa7 &adc1_in2_pc1 &adc1_in1_pc0>;
 	pinctrl-names = "default", "sleep", "arduino";
 	status = "okay";
@@ -193,7 +192,7 @@
 		status = "okay";
 		zephyr,deferred-init;
 		pinctrl-0 = <>;
-		pinctrl-1 = <&analog_pb2>;
+		pinctrl-1 = <>;
 		pinctrl-2 = <&tim8_ch4n_pb2>;
 		pinctrl-names = "default", "sleep", "arduino";
 	};
